requirements.txt
flask
flask-socketio
redis
pymongo
gunicorn
eventlet
watchgod
flask-cors
python-dotenv